backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<4.1.0
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
